Core/Scripts: converted Argent Tournament pennants to AI
Closes #14616

(cherrypicked from 804257b99a26ca2dc1941a7577e5ae06c954f9cb)
